+++ src/helpers/codehelpers.cpp
@@ -1,4 +1,5 @@
 #include "helpers/codehelpers.h"
+#include <fstream>
 #ifdef _WIN32
 #include <windows.h>
 #elif defined(__linux__)
@@ -24,5 +25,36 @@ namespace Nickvision::Helpers
 #elif defined(__linux__)
         return std::string(strerror(errno));
 #endif
+    }
+
+    std::vector<std::byte> CodeHelpers::readFileBytes(const std::filesystem::path& path)
+    {
+        if(!std::filesystem::exists(path))
+        {
+            return {};
+        }
+        std::ifstream file{ path, std::ios::binary };
+        if(file.good())
+        {
+            std::vector<std::byte> bytes(std::filesystem::file_size(path));
+            file.read(reinterpret_cast<char*>(bytes.data()), static_cast<std::streamsize>(bytes.size()));
+            return bytes;
+        }
+        return {};
+    }
+
+    bool CodeHelpers::writeFileBytes(const std::filesystem::path& path, const std::vector<std::byte>& bytes, bool overwrite)
+    {
+        if(std::filesystem::exists(path) && !overwrite)
+        {
+            return false;
+        }
+        std::ofstream file{ path, std::ios::binary | std::ios::trunc };
+        if(file.good())
+        {
+            file.write(reinterpret_cast<const char*>(bytes.data()), static_cast<std::streamsize>(bytes.size()));
+            return true;
+        }
+        return false;
     }
 }